requirement.txt
streamlit
yfinance
python-dotenv
fpdf
langchain
openai
plotly